requirements.txt
flask
flask_cors
flask_sqlalchemy
requests
python-dotenv